+++ src/components/PublicRecipeGallery/index.ts
@@ -1,0 +1,1 @@
+export { PublicRecipeGallery } from "./PublicRecipeGallery";
