+++ src/components/index.ts
@@ -4,3 +4,4 @@ export * from "./FallbackScreens";
 export * from "./RecipeCard";
 export * from "./TagPicker";
 export * from "./Animation";
+export * from "./PublicRecipeGallery";
